app/requirements.txt
Flask
mysql-connector
gunicorn